Revert "Merge branch 'main' into julian-demo2"
This reverts commit 12671a3ecf21886715f8b74f50ebae17e2f8fcba, reversing
changes made to 46e037b7056c6368db0bc49c8de8f8c255d0d39b.
