+++ .gitignore
@@ -38,5 +38,5 @@ next-env.d.ts
 node_modules/
 *.node
 
+node_modules 2/
 
-node_modules 2
