+++ package.json
@@ -10,8 +10,6 @@
     "restart": "node app.js"
   },
   "dependencies": {
-    "axios": "^1.7.7",
-    "babel-plugin-styled-components": "^2.1.4",
     "bootstrap": "^5.3.3",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
@@ -23,10 +21,9 @@
     "react-bootstrap": "^2.10.4",
     "react-dom": "^18.3.1",
     "react-image-gallery": "^1.3.0",
-    "react-router-dom": "^6.27.0",
+    "react-router-dom": "^6.26.2",
     "react-slick": "^0.30.2",
-    "slick-carousel": "^1.8.1",
-    "styled-components": "^6.1.13"
+    "slick-carousel": "^1.8.1"
   },
   "devDependencies": {
     "eslint": "^8",
