--- src/app/forgotpw/forgotpw.css
@@ -1,44 +1,0 @@
-.forgotPwBox {
-  display: flex;
-  flex-direction: column;     
-  min-height: 100vh;
-  justify-self: center;
-  align-items: center;
- 
-}
-
-.forgotPass {
-    display: flex;
-    justify-content: center;
-    border-radius: 25px;
-    background-color: rgb(107, 112, 116);
-    flex-direction: column;
-    margin: 45px auto;
-    padding: 20px;
-    align-items: center;
-    width: 100%; /* Optional: Makes the container fill horizontally */
-    max-width: 500px;
-}
-
- 
-
-#emailBox {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 100%; 
-    padding: 10px; 
-    border-radius: 5px; 
-    box-sizing: border-box; 
-}
-
-#emailBox input {
-    width: 100%;
-    padding: 10px;
-    margin-top: 10px;
-    border-radius: 5px;
-    box-sizing: border-box; 
-  }
-
-
-
