+++ src/app/login/page.js
@@ -3,20 +3,14 @@
 import Image from "next/image";
 import { useState } from 'react';
 
-
 import Header from '@/components/Header';
 import Footer from '@/components/Footer';
 import { Button } from 'react-bootstrap';
+import {Link} from 'react-router-dom';
+import ForgotPassword from "./forgotPass";
 
+import "./login.css";
 
-import Link from 'next/link';
-
-
-import styles from "./login.css";
-
-<Link href="/about" prefetch={false}>
-  About
-</Link>
 
 function Login() {
     const password = "password";
@@ -24,9 +18,10 @@ function Login() {
     const [authorized, setAuthourized] = useState(false);
     const [isForgotPassword, setIsForgotPassword] = useState(false)
 
-  
+    const handleForgotPassword = () => {
+        setIsForgotPassword(true);
+    };
     function handleSubmit(e) {
-        e.preventDefault();
         const passwordInput = e.target.querySelector(
             'input[type="password"]'
         ).value;
@@ -34,43 +29,40 @@ function Login() {
         setAuthourized(auth);
 
     }
-   
+
     const login = (
        <div className="loginBox"> 
             <h1>Login</h1>
-            <form action="#" onSubmit={handleSubmit} >
-                <div className="inputBoxes">
-                    <input type = "email" placeholder = "Email" />
-                </div>
+            <form action="#" onSubmit={handleSubmit}>
+                <input type = "email" placeholder = "Email" />
                 <input type = "password" placeholder="Password" />
                 <div className= "rmmrMe"> 
-                    <input type="checkbox" id = "checkBox"/>
                     <p> Remember Me</p>
-                    <Link href="/signup"> Sign-up</Link>
+                    <input type="checkbox" />
                 </div>
-                
-                <div className="bottom">
-                    
-                    <Button variant="primary" type="submit">Login</Button>
-                    <Link href="/forgotpw"> Forgot Password?</Link>
+                <div className="forgotPass">
+                    <button onClick={handleForgotPassword}>Forgot Password?</button>
                 </div>
-        
+
                 
+                <Button variant="primary" type="submit">Login</Button>
 
             </form>
         </div>
     );
-    
 
     return (
-        <div className="container34">
-            <Header />
-            <div id="authorized">
-                {authorized ? "LOGIN SUCCESSFUL" : login}
-            </div>  
-            <Footer/>           
+        <div>
+            {isForgotPassword ? (
+                <ForgotPassword/>
+            ) : (
+                <div id="authorized">
+                    {authorized ? "LOGIN SUCCESSFUL" : login}
+                </div>             
+            )}
+
         </div>
     );
 
-}
+};
 export default Login;
