+++ src/app/signup/page.js
@@ -1,15 +1,10 @@
 'use client'
-import styles from './signup.css'
+import './signup.css'
 import { useState } from 'react';
 import Link from 'next/link';
-import Header from '@/components/Header';
-import Footer from '@/components/Footer';
-import { Button } from 'react-bootstrap';
+
 export default function Signup() {
 
-    <Link href="/about" prefetch={false}>
-  About
-</Link>
 
     const [registered, setRegistration] = useState(false);
     const handleRegistration = () => {
@@ -17,48 +12,41 @@ export default function Signup() {
     }
 
     return (
-        <div>
-            <Header/>
-            <div className="signUpWrapper">
-                
-                {registered ? ( 
-                    <>
-                        <h1>Registration Complete!</h1>
-                        <p>Click Below to Return to the Home Page</p>
-                        <Link href="/">Home Page</Link>
-                    </>
-                ) : (
+        <div className="signUpWrapper">
+            {registered ? ( 
                 <>
-                    <h1> Sign Up</h1>
-                    <form className="formBox">
-                        <div className="inputWrapper">
-                            <label>Name:</label>
-                            <input type="text" name="name" placeholder="Name"/>
-                        </div>
-                        <div className="inputWrapper">
-                            <label>Email: </label>
-                            <input type="email" name="email" placeholder="Email"/>
-                        </div>
-                        <div className="inputWrapper">
-                            <label>Phone Number: </label>
-                            <input type="phoneNumber" name="phoneNumber" placeholder="###-###-###"/>
-                        </div>
-                        <div className="inputWrapper">
-                            <label>Password:</label>
-                            <input type="password" name="password" placeholder="Password"/>
-                        </div>
-                        <div className="inputWrapper">
-                            <label>Confirm Password: </label>
-                            <input type="password" name="confirmPass" placeholder="Confirm Password"/>
-                        </div>
-                        <button onClick={setRegistration} className="button">Submit</button>
-                    </form>
+                    <h1>Registration Complete!</h1>
+                    <p>Click Below to Return to the Home Page</p>
+                    <Link href="/">Home Page</Link>
                 </>
-                )}
-            
-            </div>
-            <Footer/>
+            ) : (
+            <>
+                <h1> Sign Up</h1>
+                <form className="formBox">
+                    <div className="inputWrapper">
+                        <label>Name:</label>
+                        <input type="text" name="name"/>
+                    </div>
+                    <div className="inputWrapper">
+                        <label>Email: </label>
+                        <input type="email" name="email"/>
+                    </div>
+                    <div className="inputWrapper">
+                        <label>Phone Number: </label>
+                        <input type="phoneNumber" name="phoneNumber"/>
+                    </div>
+                    <div className="inputWrapper">
+                        <label>Password:</label>
+                        <input type="password" name="password"/>
+                    </div>
+                    <div className="inputWrapper">
+                        <label>Confirm Password: </label>
+                        <input type="password" name="confirmPass"/>
+                    </div>
+                    <button onClick={setRegistration} className="button">Submit</button>
+                </form>
+            </>
+            )}
         </div>
     )
-   
 };
